fix rename component file to match with component name
