+++ packages/block-library/src/table/components/table-dimensions-preview.js
